+++ Raytracer/src/Raytracer/Sampling/Sampler.java
@@ -3,7 +3,7 @@ package Raytracer.Sampling;
 import Raytracer.Math.Vec2;
 
 public interface Sampler {
-	public Vec2 getModXY(int x, int y);
+	public Vec2 getModXY(double x, double y);
 	
 	public int getSamples();
 	public void setSamples(int samples);
